requirements.txt
streamlit
google-generativeai
requests
feedparser
python-dotenv
